+++ components/helpful_hours/DropDownTab.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
-import { View, FlatList, Text, StyleSheet } from "react-native";
-import { Card, List, Title } from "react-native-paper";
-import HelpfulTime from "./HelpfulTime";
+import { View, StyleSheet } from "react-native";
+import { List } from "react-native-paper";
+import AvailableHours from "./AvailableHours";
 import ContactInformation from "./ContactInformation";
 
 const DropDownTab = ({ section }) => {
@@ -28,22 +28,7 @@ const DropDownTab = ({ section }) => {
                 onPress={handlePress} //alarm-multiple or bell-ring-outline or clock-alert-outline or clock-fast
                 left={(props) => <List.Icon {...props} icon="hours-24" />}
             >
-                <Card style={styles.card}>
-                    <Title style={styles.title}>Hours</Title>
-                    {section.location && (
-                        <Text style={styles.location}>
-                            ({section.location})
-                        </Text>
-                    )}
-                    <FlatList
-                        style={{ marginBottom: 5 }}
-                        showsVerticalScrollIndicator={false}
-                        listKey={(_, index) => `_key${index.toString()}`}
-                        keyExtractor={(_, index) => `_key${index.toString()}`}
-                        data={openHours}
-                        renderItem={({ item }) => <HelpfulTime item={item} />}
-                    />
-                </Card>
+                <AvailableHours openHours={openHours} section={section} />
                 <ContactInformation contactInfo={contactInfo} />
             </List.Accordion>
         </View>
@@ -62,22 +47,6 @@ const styles = StyleSheet.create({
         borderRadius: 5,
         overflow: "hidden",
         paddingLeft: 0,
-    },
-    card: {
-        marginTop: 5,
-        alignItems: "center",
-        marginBottom: 5,
-        marginLeft: -64,
-    },
-    title: {
-        textAlign: "center",
-        fontSize: 18,
-        fontWeight: "bold",
-        paddingTop: 5,
-    },
-    location: {
-        textAlign: "center",
-        fontStyle: "italic",
     },
 });
